Removed call_loopy dependency
